Add support for CLOUDINARY_URL
This makes config a bit better: It can now *either* accept a cloud_name
or just use the CLOUDINARY_URL envvar. Currently nothing this plugin
does *needs* full secure API access, but this will make it easier to
either add that in the future, or allow the plugin to share
authentication w/ more advanced Cloudinary interaction in the same
project.